Clean up terminology regarding current/running/active tx:
We can be *in a transaction* that has been killed so is no longer *running*. Previously (also in Clojure), getRunning would return the transaction we're in, and isRunning would return true when we're in a transaction, even if that transaction was killed. Now, getCurrent/isCurrent looks at the transaction we're in; while isNotKilled looks at whether it's still running or not.